[resume-component-refactor] moving resumeData to data folder
