+++ package.json
@@ -51,6 +51,7 @@
     ]
   },
   "devDependencies": {
-    "gh-pages": "^6.1.0"
+    "gh-pages": "^6.1.0",
+    "@babel/plugin-proposal-private-property-in-object": "^7.16.7"
   }
 }
